+++ src/app/Unit-test/advanced-search-repo.spec.ts
@@ -1,0 +1,7 @@
+import { AdvancedSearchRepo } from '../Interface/advanced-search-repo';
+
+describe('AdvancedSearchRepo', () => {
+  it('should create an instance', () => {
+    expect(new AdvancedSearchRepo()).toBeTruthy();
+  });
+});