+++ src/app/Unit-test/advanced-search-user.spec.ts
@@ -1,0 +1,7 @@
+import { AdvancedSearchUser } from '../Interface/advanced-search-user';
+
+describe('AdvancedSearchUser', () => {
+  it('should create an instance', () => {
+    expect(new AdvancedSearchUser()).toBeTruthy();
+  });
+});