+++ src/app/Unit-test/git-search-user.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { GitSearchUserService } from './git-search-user.service';
+import { GitSearchUserService } from '../Service/git-search-user.service';
 
 describe('GitSearchUserService', () => {
   let service: GitSearchUserService;