+++ src/app/app-routing.module.ts
@@ -1,22 +1,22 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { GitSearchTopicComponent } from './git-search-topic/git-search-topic.component';
-import { GitSearchUserComponent } from './git-search-user/git-search-user.component';
-import { HomepageComponent } from './homepage/homepage.component';
-import { NotFoundComponent } from './not-found/not-found.component';
+import { GitSearchRepoComponent } from './Component/git-search-repo/git-search-repo.component';
+import { GitSearchUserComponent } from './Component/git-search-user/git-search-user.component';
+import { HomepageComponent } from './Component/homepage/homepage.component';
+import { NotFoundComponent } from './Component/not-found/not-found.component';
 
 const routes: Routes = [
   { path: '',
     component: HomepageComponent
   },
-  { path: 'search-topic',
-    redirectTo: '/search-topic/angular',
+  { path: 'search-repo',
+    redirectTo: '/search-repo/angular',
     pathMatch: 'full'
   },
-  { path: 'search-topic/:query',
-    component: GitSearchTopicComponent,
+  { path: 'search-repo/:query',
+    component: GitSearchRepoComponent,
     data: {
-      title: 'Git Search Topic'
+      title: 'Git Search Repositories'
     }
   },
   { path: 'search-user',  
@@ -26,7 +26,7 @@ const routes: Routes = [
   { path: 'search-user/:query',  
     component: GitSearchUserComponent,
     data: {
-      title: 'Git Search User'
+      title: 'Git Search Users'
     }
   },
   { path: '**', component: NotFoundComponent }