+++ src/app/app.module.ts
@@ -7,31 +7,43 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 // Service
-import { GitSearchUserService } from './git-search-user.service';
-import { GitSearchTopicService } from './git-search-topic.service';
+import { GitSearchRepoService } from './Service/git-search-repo.service';
+import { GitSearchUserService } from './Service/git-search-user.service';
+import { GitSearchCodeService } from './Service/git-search-code.service';
+import { UnifiedSearchService } from './Service/unified-search.service';
 // Component
-import { GitSearchTopicComponent } from './git-search-topic/git-search-topic.component';
-import { GitSearchUserComponent } from './git-search-user/git-search-user.component';
-import { HomepageComponent } from './homepage/homepage.component';
-import { NotFoundComponent } from './not-found/not-found.component';
+import { GitSearchRepoComponent } from './Component/git-search-repo/git-search-repo.component';
+import { GitSearchUserComponent } from './Component/git-search-user/git-search-user.component';
+import { HomepageComponent } from './Component/homepage/homepage.component';
+import { NotFoundComponent } from './Component/not-found/not-found.component';
+import { ReposDisplayComponent } from './Advanced-Component/repos-display/repos-display.component';
+import { CodesDisplayComponent } from './Advanced-Component/codes-display/codes-display.component';
 
 
 @NgModule({
   declarations: [
     AppComponent,
-    GitSearchTopicComponent,
+    GitSearchRepoComponent,
     GitSearchUserComponent,
     HomepageComponent,
-    NotFoundComponent
+    NotFoundComponent,
+    ReposDisplayComponent,
+    CodesDisplayComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
     HttpClientModule,
     AppRoutingModule,
+    ReactiveFormsModule,
     BrowserAnimationsModule
   ],
-  providers: [GitSearchTopicService, GitSearchUserService],
+  providers: [
+    GitSearchRepoService, 
+    GitSearchUserService, 
+    GitSearchCodeService,
+    UnifiedSearchService
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }