+++ src/app/interface/advanced-search-repo.ts
@@ -1,0 +1,10 @@
+export class AdvancedSearchRepo {
+    constructor(
+        public repository: string,
+        public language?: string,
+        public user?: string,
+        public size?: number,
+        public stars?: number,
+        public topic?: string,
+    ) {}
+}