+++ src/app/interface/advanced-search-user.ts
@@ -1,0 +1,9 @@
+export class AdvancedSearchUser {
+    constructor(
+        public user: string,
+        public language?: string,
+        public repos?: number,
+        public followers?: number,
+        public location?: string
+    ) {}
+}