+++ src/app/interface/unified-search.ts
@@ -1,0 +1,7 @@
+import { GitSearchRepo } from './git-search-repo';
+import { GitSearchCode } from './git-search-code';
+
+export interface UnifiedSearch {
+    repositories: GitSearchRepo,
+    code: GitSearchCode
+}