+++ record/Changelog.md
@@ -1,6 +1,24 @@
 ## Github Search Angular Web Application
+<a name="0.0.4"></a>
+# 0.0.4 (2020-07-29)
+* Create unified search service
+    * Display both get('repositories') and get('code')
+    * Merge reponse by RxJS forkJoin
+* Create Advanced-Component folder for display multiple component
+    * Create repos display component
+    * Create codes display component
+* Create addFavorite and deFavorite in git-search-repo
+
+<a name="0.0.3"></a>
+# 0.0.3 (2020-07-25)
+* Update all foler from topic to repo
+* Add repo search reactive form and template-driven form
+    * Add validation to check for special character
+* Add repo user reactive form and template-driven form
+    * Add validation to check for special character
+
 <a name="0.0.2"></a>
-# 0.0.1 (2020-07-24)
+# 0.0.2 (2020-07-24)
 * Remove @fortawesome package
     * Add font awesome 6 link in index.html